f2l-web-refactor/backend/requirements.txt
# Core Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Task Queue
celery==5.3.6
redis==5.0.1
flower==2.0.1

# AWS S3
boto3==1.34.27
botocore==1.34.27

# FTP/SFTP
paramiko==3.4.0

# WebSocket
python-socketio==5.11.0
python-engineio==4.9.0

# Authentication & Security
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client
httpx==0.26.0
aiofiles==23.2.1

# Utilities
python-dateutil==2.8.2
pytz==2024.1
tenacity==8.2.3

# Monitoring & Logging
prometheus-client==0.19.0
psutil==5.9.8

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0
